{{cookiecutter.project_name}}/requirements.txt
Django==1.11.16
django-filter==1.0.4
djangorestframework==3.9.0
djangorestframework-httpsignature==1.0.0
httpsig==1.3.0
MySQL-python==1.2.5
pycryptodome==3.8.1
pytz==2018.9
six==1.12.0
django-cors-headers==2.5.2
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.0
enum34==1.1.6
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.4.1
scandir==1.10.0
simplegeneric==0.8.1
traitlets==4.3.2
wcwidth==0.1.7
